pyproject.toml
[tool.poetry]
name = "ec2-existing-vpc-dynamo-configuration"
version = "0.1.0"
description = "Deploy ec2 instances to existing vpc based on confirguation in dyanmodb"
authors = ["Galen Dunkleberger <awsgalen@amazon.com>"]


[tool.poetry.dependencies]
python="^3.9"
"aws-cdk.core" = "^1.134.0"
"aws-cdk.aws-iam" = "^1.134.0"
"aws-cdk.aws-ec2" = "^1.134.0"
boto3 = "^1.19.4"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"





[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
